[strings/fence] Заборчик #219
Fix compiler warning

+++ problems/strings/fence/solution.cpp
@@ -7,7 +7,7 @@ int main()
   std::getline(std::cin, input);
 
   int counter = 0;
-  for (int i = 0; i < input.size(); ++i)
+  for (size_t i = 0; i < input.size(); ++i)
   {
     if (input.at(i) == ' ')
       continue;
